Extract equation_symbol_mapping to its own files
Because we can.
